fix normalization issues in G2lat.H2ThPh to avoid divide by zeros spin RB GUI mods add contour plot of sqrt(W)*delt-Y plot fix np.int -> np.int32 in G2spc & shapes
git-svn-id: https://subversion.xray.aps.anl.gov/pyGSAS/trunk@5513 2b3fe19a-dc8f-cc80-9b5f-a4fb0fcdb5be

+++ G2shapes.py
@@ -1096,7 +1096,7 @@ def G2shapes(Profile,ProfDict,Limits,data,dlg=None):
         xyz = np.array([x1,y1,z1])
         drs = get_drs(xyz,XYZ)
         drs_grid = np.where(drs>max_dr,max_dr,drs)/hist_grid
-        int_drs_grid = np.array(drs_grid,dtype=np.int)
+        int_drs_grid = np.array(drs_grid,dtype=np.int32)
         int_drs_grid = np.where(int_drs_grid>num_hist-2,num_hist-2,int_drs_grid)                
         ip_lows = int_drs_grid
         ip_highs = ip_lows + 1
